requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
python-docx==1.1.0
groq==0.4.2